[curve25519] Cleanup public API
