+++ cryptography/curve25519/benches/batch_verify.rs
@@ -1,21 +1,20 @@
-//! Throughput of [`verify_batch_bytes`] across batch sizes, up to the ~16k-signature scale a busy
-//! validator sees per block, starting from raw wire bytes for everything (public key, signature,
-//! message) -- so this measures decompressing both `A` and `R` plus the MSM, not just the MSM.
-//! Each signature is from an independent key over an independent message (no key/message reuse to
-//! amortize), the harder case for the underlying MSM.
-use commonware_cryptography_curve25519::signing::{Signature, verify_batch_bytes};
+//! Throughput of the public batch-verification API across batch sizes. Each signature is from an
+//! independent key over an independent message (no key/message reuse to amortize), the harder case
+//! for the underlying MSM. Fixture generation and verifier construction are not timed.
+use commonware_cryptography_curve25519::signing::{
+    BatchVerifier, Signature, SigningKey, VerifyingKey,
+};
+use commonware_math::algebra::Random;
 use commonware_parallel::{Rayon, Sequential};
 use commonware_utils::{NZUsize, TestRng, test_rng};
 use criterion::{BatchSize, Criterion, Throughput, criterion_group, criterion_main};
-use ed25519_consensus::SigningKey as RefSigningKey;
-use rand_core::Rng;
 use std::hint::black_box;
 
-/// Generates `n` valid `(pubkey_bytes, Signature, message)` triples, signed by independent keys
-/// over independent messages using the `ed25519-consensus` reference implementation (this crate
-/// does not yet implement signing). Nothing here is decompressed: that is the point being
-/// measured.
-fn generate_batch(n: usize) -> Vec<([u8; 32], Signature, Vec<u8>)> {
+const NAMESPACE: &[u8] = b"_COMMONWARE_CRYPTOGRAPHY_CURVE25519_BATCH_VERIFY_BENCH";
+
+/// Generates `n` valid `(verifying key, signature, message)` triples, signed by independent keys
+/// over independent messages using the public API.
+fn generate_batch(n: usize) -> Vec<(VerifyingKey, Signature, Vec<u8>)> {
     generate_batch_with_message_size(n, 0)
 }
 
@@ -24,14 +23,11 @@ fn generate_batch(n: usize) -> Vec<([u8; 32], Signature, Vec<u8>)> {
 fn generate_batch_with_message_size(
     n: usize,
     message_size: usize,
-) -> Vec<([u8; 32], Signature, Vec<u8>)> {
+) -> Vec<(VerifyingKey, Signature, Vec<u8>)> {
     let mut rng = test_rng();
     (0..n)
         .map(|i| {
-            let mut seed = [0u8; 32];
-            rng.fill_bytes(&mut seed);
-            let signing_key = RefSigningKey::from(seed);
-            let pubkey_bytes = signing_key.verification_key().to_bytes();
+            let signing_key = SigningKey::random(&mut rng);
 
             let message = if message_size == 0 {
                 format!("message {i}").into_bytes()
@@ -40,19 +36,27 @@ fn generate_batch_with_message_size(
                 message[..8].copy_from_slice(&(i as u64).to_le_bytes());
                 message
             };
-            let signature = Signature::from_bytes(signing_key.sign(&message).to_bytes());
+            let signature = signing_key.sign(NAMESPACE, &message);
 
-            (pubkey_bytes, signature, message)
+            (signing_key.verifying_key(), signature, message)
         })
         .collect()
+}
+
+fn batch_verifier(batch: &[(VerifyingKey, Signature, Vec<u8>)]) -> BatchVerifier {
+    let mut verifier = BatchVerifier::new(batch.len());
+    for (verifying_key, signature, message) in batch {
+        verifier.add(NAMESPACE, message, verifying_key, signature);
+    }
+    verifier
 }
 
 /// Nested so `module_path!()` includes the crate::module separator the benchmark-name lint (and
 /// the benchmark-tracking dashboard) expect.
 mod verify_batch_bytes_bench {
     use super::{
-        BatchSize, Criterion, NZUsize, Rayon, Sequential, TestRng, Throughput, black_box,
-        generate_batch, verify_batch_bytes,
+        BatchSize, Criterion, NZUsize, Rayon, Sequential, TestRng, Throughput, batch_verifier,
+        black_box, generate_batch,
     };
 
     pub fn bench(c: &mut Criterion) {
@@ -67,15 +71,13 @@ mod verify_batch_bytes_bench {
                 group.throughput(Throughput::Elements(n as u64));
                 group.bench_function(format!("msg1232/sigs={n} conc={concurrency}"), |b| {
                     b.iter_batched(
-                        || TestRng::new(1),
-                        |mut rng| {
-                            let items =
-                                batch.iter().map(|(pk, sig, msg)| (pk, sig, msg.as_slice()));
+                        || (TestRng::new(1), batch_verifier(&batch)),
+                        |(mut rng, verifier)| {
                             #[allow(clippy::option_if_let_else)]
                             if let Some(rayon) = rayon.as_ref() {
-                                black_box(verify_batch_bytes(&mut rng, items, rayon))
+                                black_box(verifier.verify(&mut rng, rayon))
                             } else {
-                                black_box(verify_batch_bytes(&mut rng, items, &Sequential))
+                                black_box(verifier.verify(&mut rng, &Sequential))
                             }
                         },
                         BatchSize::SmallInput,
@@ -90,15 +92,13 @@ mod verify_batch_bytes_bench {
                 group.throughput(Throughput::Elements(n as u64));
                 group.bench_function(format!("sigs={n} conc={concurrency}"), |b| {
                     b.iter_batched(
-                        || TestRng::new(1),
-                        |mut rng| {
-                            let items =
-                                batch.iter().map(|(pk, sig, msg)| (pk, sig, msg.as_slice()));
+                        || (TestRng::new(1), batch_verifier(&batch)),
+                        |(mut rng, verifier)| {
                             #[allow(clippy::option_if_let_else)]
                             if let Some(rayon) = rayon.as_ref() {
-                                black_box(verify_batch_bytes(&mut rng, items, rayon))
+                                black_box(verifier.verify(&mut rng, rayon))
                             } else {
-                                black_box(verify_batch_bytes(&mut rng, items, &Sequential))
+                                black_box(verifier.verify(&mut rng, &Sequential))
                             }
                         },
                         BatchSize::SmallInput,
